refactor(main): extract password-store args, mpv plugin detection, yomitan anki sync, session bindings, log export from main.ts
